Обновление конфигурации и рефакторинг кода
Основные изменения:
- Добавлена переменная IS_DEBUG в файл .env.example для управления режимом отладки.
- Упрощена и улучшена структура кода, включая рефакторинг функций получения и обработки временных меток.
- Изменены методы обработки ошибок для более ясного и понятного кода.
- Обновлены импорты для улучшения читаемости и структуры проекта.

+++ src-tauri/src/modules/events/record.rs
@@ -23,11 +23,11 @@ pub fn set_event_channel_record_global(channel: Channel<RecordEvent>) {
 #[serde(rename_all = "camelCase", tag = "event", content = "data")]
 pub enum RecordEvent {
     #[serde(rename_all = "camelCase")]
-    Start { timestamp: u128 },
+    Start { timestamp: u64 },
     #[serde(rename_all = "camelCase")]
-    Progress { timestamp: u128, peak: SampleType },
+    Progress { timestamp: u64, peak: SampleType },
     #[serde(rename_all = "camelCase")]
-    Stop { timestamp: u128 },
+    Stop { timestamp: u64 },
 }
 
 impl RecordEvent {
